+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Update dependencies to resolve security vulnerabilities
 
+### Changed
+- Update noble to latest version (2.3.4)
+
 ## [1.2.5] - 2025-08-01
 
 ### Fixed